Add python bindings for SIOLegacyReader
