+++ python/podio/sio_io.py
@@ -23,3 +23,22 @@ class Reader(BaseReaderMixin):
     self._reader.openFile(filename)
 
     super().__init__()
+
+
+class LegacyReader(BaseReaderMixin):
+  """Reader class for reading legcy podio sio files.
+
+  This reader can be used to read files that have not yet been written using the
+  Frame based I/O into Frames for a more seamless transition.
+  """
+
+  def __init__(self, filename):
+    """Create a reader that reads from the passed file.
+
+    Args:
+        filename (str): File to open and read data from
+    """
+    self._reader = podio.SIOLegacyReader()
+    self._reader.openFile(filename)
+
+    super().__init__()
