+++ python/podio/test_ReaderRoot.py
@@ -4,7 +4,7 @@
 import unittest
 
 from podio.root_io import Reader, LegacyReader
-from podio.test_Reader import ReaderTestCaseMixin
+from podio.test_Reader import ReaderTestCaseMixin, LegacyReaderTestCaseMixin
 
 
 class RootReaderTestCase(ReaderTestCaseMixin, unittest.TestCase):
@@ -14,34 +14,8 @@ class RootReaderTestCase(ReaderTestCaseMixin, unittest.TestCase):
     self.reader = Reader('example_frame.root')
 
 
-class RootLegacyReaderTestCase(unittest.TestCase):
-  """Test cases for the legacy root input files and reader.
-
-  NOTE: Since also the LegacyReader uses the BaseReaderMixin many invalid access
-  test cases are already covered by the ReaderTestCaseMixin and here we focus on
-  the happy paths as this is slightly different.
-  """
+class RootLegacyReaderTestCase(LegacyReaderTestCaseMixin, unittest.TestCase):
+  """Test cases for the legacy root input files and reader."""
   def setUp(self):
     """Setup a reader, reading from the example files"""
     self.reader = LegacyReader('example.root')
-
-  def test_categories(self):
-    """Make sure the legacy reader returns only one category"""
-    cats = self.reader.categories
-    self.assertEqual(("events",), cats)
-
-  def test_frame_iterator(self):
-    """Make sure the FrameIterator works."""
-    frames = self.reader.get('events')
-    self.assertEqual(len(frames), 2000)
-
-    for i, frame in enumerate(frames):
-      # Rudimentary check here only to see whether we got the right frame
-      self.assertEqual(frame.get_parameter('UserEventName'), f' event_number_{i}')
-      # Only check a few Frames here
-      if i > 10:
-        break
-
-    # Index based access
-    frame = frames[123]
-    self.assertEqual(frame.get_parameter('UserEventName'), ' event_number_123')
